+++ PCL.Core/Minecraft/Profile/Authentication/IAuthenticateProvider.cs
@@ -1,0 +1,10 @@
+﻿using System.Threading.Tasks;
+
+namespace PCL.Core.Minecraft.Profile.Authentication;
+
+public interface IAuthenticateProvider
+{
+    //public Task AuthenticateAsync();
+
+    public void RefreshInfo();
+}
