+++ PCL.Core/Minecraft/Profile/Authentication/MicrosoftProvider.cs
@@ -1,0 +1,10 @@
+﻿using PCL.Core.IO.Net;
+using PCL.Core.Minecraft.IdentityModel.Extensions.OpenId;
+using PCL.Core.Minecraft.IdentityModel.OAuth;
+
+namespace PCL.Core.Minecraft.Profile.Authentication;
+
+public class MicrosoftProvider(OpenIdClient client)
+{
+    
+}
