+++ PCL.Core/Minecraft/Profile/Authentication/Utils/Models/XboxLiveResponse.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace PCL.Core.Minecraft.Profile.Authentication.Utils.Models;
+
+public record XboxLiveResponse
+{
+    public required DateTime IssueInstant { get; set; }
+    public required DateTime NotAfter { get; set; }
+    public required string Token { get; set; }
+}
