--- PCL.Core/Minecraft/Profile/Autnentication/IAuthenticateProvider.cs
@@ -1,8 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace PCL.Core.Minecraft.Profile.Autnenrication;
-
-public interface IAuthenticateProvider
-{
-    //public Task AuthenticateAsync();
-}
