--- PCL.Core/Minecraft/Profile/Autnentication/MicrosoftProvider.cs
@@ -1,10 +1,0 @@
-﻿using PCL.Core.IO.Net;
-using PCL.Core.Minecraft.IdentityModel.Extensions.OpenId;
-using PCL.Core.Minecraft.IdentityModel.OAuth;
-
-namespace PCL.Core.Minecraft.Profile.Autnenrication;
-
-public class MicrosoftProvider(OpenIdClient client)
-{
-    
-}
