--- PCL.Core/Minecraft/Profile/Autnentication/Utils/Models/MojangIdentityToken.cs
@@ -1,9 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace PCL.Core.Minecraft.Profile.Autnenrication.Utils.Models;
-
-public class MojangIdentityToken
-{
-    [JsonPropertyName("identityToken")]
-    public required string IdentityToken { get; set; }
-}
