--- PCL.Core/Minecraft/Profile/Autnentication/Utils/Models/XboxAuthenticate.cs
@@ -1,8 +1,0 @@
-﻿namespace PCL.Core.Minecraft.Profile.Autnenrication.Utils.Models;
-
-public record XboxAuthenticate<T>
-{
-    public required T Properties { get; set; }
-    public string RelyingParty { get; set; } = "http://auth.xboxlive.com";
-    public required string TokenType { get; set; }
-}
