+++ PCL.Core/Minecraft/Profile/Models/McProfile.cs
@@ -1,0 +1,12 @@
+﻿using PCL.Core.Minecraft.Profile.Autnenrication;
+
+namespace PCL.Core.Minecraft.Profile.Models;
+
+internal record McProfile: SafeProfile
+{
+    
+    public static IAuthenticateProvider? CreateAuthenticateServiceProvider()
+    {
+        return default;
+    }
+}
