+++ PCL.Core/Minecraft/Profile/Models/SafeProfile.cs
@@ -2,6 +2,9 @@
 
 namespace PCL.Core.Minecraft.Profile.Models;
 
+/// <summary>
+/// 适用于外部使用的安全档案类
+/// </summary>
 public record SafeProfile
 {
     [JsonPropertyName("username")]
@@ -12,4 +15,8 @@ public record SafeProfile
     public required string Uuid { get; set; }
     [JsonPropertyName("tokenType")]
     public required string TokenType { get; set; }
+    [JsonPropertyName("profileType")]
+    public ProfileType? ProfileType { get; set; }
+
+    private McProfile? _innerProfile;
 }
